Small restructure and image size fix
