--- src/main/java/fish/payara/hello/entities/GamesGenresBridge.java
@@ -1,46 +1,0 @@
-package fish.payara.hello.entities;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "games_genres")
-public class GamesGenresBridge {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Basic
-    @Column(name = "game_genres_id")
-    private Long id;
-
-//    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-//    @JoinColumn(name = "game_id", nullable = false)
-    private Games game;
-
-//    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-//    @JoinColumn(name = "genre_id", nullable = false)
-//    private Genres genre;
-
-//
-//    public Games getGame() {
-//        return game;
-//    }
-//
-//    public void setGame(Games game) {
-//        this.game = game;
-//    }
-//
-//    public Genres getGenre() {
-//        return genre;
-//    }
-//
-//    public void setGenre(Genres genre) {
-//        this.genre = genre;
-//    }
-//
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public Long getId() {
-        return id;
-    }
-}
